test: LetterService Test 코드 작성
